fix Coverity findings, closes #50

+++ src/misc.h
@@ -6,7 +6,7 @@
 int kerneltest(void);
 int spacecheck(char *path);
 void sighandler(int);
-int getbtime(void);
+uint64_t getbtime(void);
 char *getvalue(uint64_t mb, uint64_t kb, int len, int type);
 char *getrate(uint64_t mb, uint64_t kb, uint32_t interval, int len);
 char *gettrafficrate(uint64_t bytes, uint32_t interval, int len);
